modularization of components
