+++ src/features/InputBar.js
@@ -1,0 +1,17 @@
+import { inputId } from './dataSlice';
+import { useDispatch, useSelector } from 'react-redux';
+
+function InputBar() {
+    const dispatch = useDispatch();
+    const data = useSelector((state) => state.data);
+    return (
+        <input
+            value={data.objectId}
+            onChange={(e) => {
+                dispatch(inputId(Number(e.target.value)));
+            }}
+        />
+    );
+}
+
+export default InputBar;
